translator/requirements.txt
functions-framework==3.*
google-cloud==0.34.0
alive-progress==3.1.0